moving components files
