+++ src/components/utils/forms/FilterForm.js
@@ -1,7 +1,7 @@
 /* eslint-disable react-hooks/exhaustive-deps */
 /* eslint-disable eqeqeq */
 import React, { useState } from 'react'
-import api from '../../services/GeneralApi'
+import api from '../../../services/GeneralApi'
 
 import { InputText } from 'primereact/inputtext'
 import { Dropdown } from 'primereact/dropdown'
